Move ExecutionResult to its own module
